refactor(web): step 10 — extract chat-scroll-rail component; jump semantics stay with useChatScroll in chat-pane
